notes/app/requirements.txt
flask
Flask-RESTful
requests
flask-mysql
